chore(core): separate unit and integration tests for webmangler root
- Split the test suite for src/index.ts into a unit test (checking
  exported values) and integration tests (running _WebMangler_)
- Switch from the deprecated `TestScenario` to `TestScenarios` in the
  embeds test suite.
- Prefer using separate `TestCase` type definition in the helpers test
  suite for readability
- Combine the languages.test.ts and manglers.test.ts with the respective
  unit/index.test.ts for the languages and manglers directories. This
  includes fixing the name of the import in the unit tests for languages
