--- packages/core/src/__tests__/langauges.test.ts
@@ -1,21 +1,0 @@
-import { expect } from "chai";
-
-import * as languages from "../languages";
-
-suite("Exports of webmangler/languages", function() {
-  test("has BuiltInLanguagesSupport", function() {
-    expect(languages).to.haveOwnProperty("BuiltInLanguagesSupport");
-  });
-
-  test("has CssLanguageSupport", function() {
-    expect(languages).to.haveOwnProperty("CssLanguageSupport");
-  });
-
-  test("has HtmlLanguageSupport", function() {
-    expect(languages).to.haveOwnProperty("HtmlLanguageSupport");
-  });
-
-  test("has JavaScriptLanguageSupport", function() {
-    expect(languages).to.haveOwnProperty("JavaScriptLanguageSupport");
-  });
-});
